requirements.txt
langgraph>=0.4
langchain-core>=0.3
langchain-openai>=0.3
openai>=1.0
discord.py>=2.3
motor>=3.3
pymongo>=4.6
python-dotenv>=1.0
numpy>=1.24